fixed json issues
