--- TutorsHubv1.1/package.json
@@ -1,14 +1,0 @@
-{
-  "name": "tutorshubv1.1",
-  "version": "1.0.0",
-  "main": "App.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "",
-  "license": "ISC",
-  "description": "",
-  "dependencies": {
-    "expo": "^51.0.37"
-  }
-}
