--- TutorsHubv1.1/styles/global.js
@@ -1,13 +1,0 @@
-import { StyleSheet } from 'react-native';
-
-export const globalStyles = StyleSheet.create({
-    container: {
-        flex: 1,
-        padding: 20,
-    },
-    title: {
-        fontSize: 18,
-        fontWeight: 'bold',
-        color: '#333',
-    },
-});
