+++ TutorsHubv1.11/app.json
@@ -1,7 +1,7 @@
 {
   "expo": {
-    "name": "hello-react-navigation",
-    "slug": "hello-react-navigation",
+    "name": "TutorsHubv1.11",
+    "slug": "TutorsHubv1.11",
     "version": "1.0.0",
     "orientation": "portrait",
     "icon": "./assets/icon.png",
